# contracts/server-management.openapi.yaml
openapi: 3.1.0
info:
  title: Avorion Web 管理中心 - 服务器管理 API
  version: 0.12.0
  description: >-
    服务器管理八个已确认页面的真实数据合同。任何不可取得的数据必须返回 null、unknown
    或明确错误，不允许使用前端默认值冒充真实结果。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /session/local:
    post:
      tags: [Security]
      operationId: createLocalAdminSession
      security: []
      description: 仅允许从 Server Agent 本机建立短期管理员会话，设置 HttpOnly、SameSite=Strict Cookie。
      responses:
        "200":
          description: 本机管理员会话与写操作防伪令牌
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminSessionStatus"}
        default: {$ref: "#/components/responses/Error"}
  /session:
    get:
      tags: [Security]
      operationId: getAdminSession
      security: []
      responses:
        "200":
          description: 当前浏览器会话状态；响应禁止缓存
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminSessionStatus"}
    delete:
      tags: [Security]
      operationId: revokeAdminSession
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      responses:
        "204": {description: 会话已撤销}
        default: {$ref: "#/components/responses/Error"}
...
  /servers/{serverId}/status:
    parameters: [{$ref: "#/components/parameters/ServerId"}]
    get:
      tags: [Servers]
      operationId: getServerStatus
      responses:
        "200":
          description: 真实服务器状态
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ServerStatus"}
        default: {$ref: "#/components/responses/Error"}
...
  /servers/{serverId}/events:
    parameters:
      - {$ref: "#/components/parameters/ServerId"}
      - name: limit
        in: query
        schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    get:
      tags: [Control]
      operationId: listServerEvents
      responses:
        "200":
          description: 控制页最近真实事件
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/ServerEvent"}}
        default: {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: avorion_admin_session
  parameters:
    ServerId:
      name: serverId
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 80}
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string}
...
  responses:
    Error:
      description: 统一错误
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
  schemas:
    AdminSessionStatus:
      type: object
      required: [authenticated, mode]
      properties:
        authenticated: {type: boolean}
        mode: {type: string, const: local-loopback}
        csrfToken: {type: [string, "null"]}
        expiresAt: {type: [string, "null"], format: date-time}
...
    Provenance:
      type: object
      required: [sampledAt, source, freshness]
      properties:
        sampledAt: {type: string, format: date-time}
        source:
          type: string
          enum: [windows, filesystem, steamcmd, steam-query, rcon, database]
        freshness: {type: string, enum: [live, recent, stale, unavailable]}
        unavailableReason: {type: [string, "null"]}
...
    Lifecycle:
      type: string
      enum: [running, stopped, starting, stopping, restarting, unknown]
    ConnectionStatus:
      type: string
      enum: [connected, disconnected, degraded, unknown]
    ServerStatus:
      type: object
      required: [serverId, name, lifecycle, agent, provenance]
      properties:
        serverId: {type: string}
        name: {type: string}
        lifecycle: {$ref: "#/components/schemas/Lifecycle"}
        processId: {type: [integer, "null"]}
        version: {type: [string, "null"]}
        galaxyName: {type: [string, "null"]}
        uptimeSeconds: {type: [integer, "null"], minimum: 0}
        onlinePlayers: {type: [integer, "null"], minimum: 0}
        maxPlayers: {type: [integer, "null"], minimum: 0}
        rcon:
          type: object
          required: [status]
          properties:
            status: {$ref: "#/components/schemas/ConnectionStatus"}
        steamQuery:
          type: object
          required: [status]
          properties:
            status: {$ref: "#/components/schemas/ConnectionStatus"}
            latencyMs: {type: [number, "null"], minimum: 0}
        lastSave:
          type: object
          properties:
            at: {type: [string, "null"], format: date-time}
            source: {type: [string, "null"], enum: [filesystem, rcon, null]}
        agent:
          type: object
          required: [connected]
          properties:
            connected: {type: boolean}
            lastHeartbeatAt: {type: [string, "null"], format: date-time}
        provenance: {$ref: "#/components/schemas/Provenance"}
    ServerEvent:
      type: object
      required: [eventId, occurredAt, kind, message, source]
      properties:
        eventId: {type: string}
        occurredAt: {type: string, format: date-time}
        kind: {type: string, enum: [player, save, rcon, warning, operation, other]}
        message: {type: string}
        source: {type: string, enum: [filesystem, rcon, database]}
...
    ErrorDetail:
      type: object
      required: [code, message, requestId, retryable]
      properties:
        code: {type: string}
        message: {type: string}
        requestId: {type: string}
        retryable: {type: boolean}
        details: {type: object, additionalProperties: true}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: {$ref: "#/components/schemas/ErrorDetail"}